try 배포

+++ client/src/components/pages/Review.js
@@ -102,8 +102,31 @@ export default function Review({ boardId, productMemberId }) {
     }
   };
 
-  const handleReviewButtonClick = () => {
-    setIsReviewFormVisible(!isReviewFormVisible);
+  const handleReviewButtonClick = async() => {
+    if (!memberId) {
+      // 윤혜님 여기요
+      alert("로그인이 필요한 서비스입니다.");
+      return;
+    }
+    try {
+      // 권한 체크 요청
+      const response = await axios.get(
+        `${process.env.REACT_APP_DB_HOST}review/check_authority`,
+        { params: { memberId: memberId, boardId: boardId } }
+      );
+
+      if (response.status === 200) {
+        // 권한이 있으면 리뷰 작성 폼 표시
+        setIsReviewFormVisible(!isReviewFormVisible);
+      } else if (response.status === 403) {
+        // 윤혜님 여기요
+        alert(response.data);
+      }
+    } catch (error) {
+      // 윤혜님 여기요
+      alert(error.response.data);
+      console.log(error);
+    }
   };
 
   const handleReviewUpdate = async (commentId, reviewData) => {
@@ -124,17 +147,20 @@ export default function Review({ boardId, productMemberId }) {
   };
 
   const handleReviewDelete = async (commentId) => {
-    try {
-      const response = await axios.delete(
-        `${process.env.REACT_APP_DB_HOST}review/delete/${commentId}`
-      );
-      if (response.status === 200) {
-        setModalType('리뷰가 성공적으로 삭제되었습니다.');
-        onEditToggle();
-        getReviews();
+    // 윤혜님 여기요
+    if (window.confirm('후기를 삭제하면 다시 작성할 수 없습니다.')) {
+      try {
+        const response = await axios.delete(
+          `${process.env.REACT_APP_DB_HOST}review/delete/${commentId}`
+        );
+        if (response.status === 200) {
+          setModalType('리뷰가 성공적으로 삭제되었습니다.');
+          onEditToggle();
+          getReviews();
+        }
+      } catch (error) {
+        setModalType('리뷰 삭제에 실패했습니다. 잠시 후 다시 시도해주세요.');
       }
-    } catch (error) {
-      setModalType('리뷰 삭제에 실패했습니다. 잠시 후 다시 시도해주세요.');
     }
   };
